+++ demo/src/main/java/com/example/demo/dtos/ItemResult.java
@@ -2,11 +2,13 @@ package com.example.demo.dtos;
 
 import com.example.demo.entities.Item;
 
+import java.math.BigDecimal;
+
 public record ItemResult(
         Long id,
         String name,
         int quantity,
-        double price
+        BigDecimal price
 ) {
     public static ItemResult from(Item item) {
         return new ItemResult(
